Made GCC happy
GCC was warning about offsets of bit fields, something changed in
version 4.4. The fix was to make sure the size of the field doesn't
cross the width of the type. uint8_t was changed to uint16_t for RType
and IType bitfields.

+++ include/mips-emulator/instruction.hpp
@@ -80,18 +80,18 @@ namespace mips_emulator {
         });
 
         PACKED(struct RType {
-            uint8_t zero : 6;
-            uint8_t rs : 5;
-            uint8_t rt : 5;
-            uint8_t rd : 5;
-            uint8_t shamt : 5;
-            uint8_t func : 6;
+            uint16_t zero : 6;
+            uint16_t rs : 5;
+            uint16_t rt : 5;
+            uint16_t rd : 5;
+            uint16_t shamt : 5;
+            uint16_t func : 6;
         });
 
         PACKED(struct IType {
-            uint8_t op : 6;
-            uint8_t rs : 5;
-            uint8_t rt : 5;
+            uint16_t op : 6;
+            uint16_t rs : 5;
+            uint16_t rt : 5;
             uint16_t imm : 16;
         });
